fix: use mcp catalog remote config form for url manifests
