+++ ui/user/src/lib/components/mcp/McpSetupWizard.svelte
@@ -133,6 +133,7 @@
 	{project}
 	{legacyBundleId}
 	prefilledConfig={projectMcpServerInfo}
+	manifestType={!!projectMcpServerInfo?.url ? 'url' : 'command'}
 	onUpdate={async (mcpServerInfo) => {
 		setup(mcpServerInfo);
 	}}
